Move each link type in it's own file
